% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{importMSAnalysis}
\alias{importMSAnalysis}
\title{Import MSA information}
\usage{
importMSAnalysis(
  metapath,
  pro = NA,
  characteristic = NA,
  prodata = NA,
  propath = NA,
  characteristicpath = NA,
  prodatapath = NA
)
}
\description{
Import MSA information
}
