% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{importProData}
\alias{importProData}
\title{Import information}
\usage{
importProData(
  datapath,
  pro = NA,
  characteristic = NA,
  propath = NA,
  characteristicpath = NA
)
}
\description{
Import information
}
